requirements.txt
aiocoap==0.4.5
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
network==0.1
ntplib==0.4.0
paho-mqtt==1.6.1
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
Werkzeug==2.2.3